Separate Min and Max tests.
