+++ src/System.Linq/tests/MinTests.cs
@@ -6,7 +6,7 @@ using Xunit;
 
 namespace System.Linq.Tests
 {
-    public class MinMaxTests
+    public class MinTests
     {
         [Fact]
         public void MinInt32()
@@ -20,20 +20,6 @@ namespace System.Linq.Tests
             Assert.Equal(100, hundred.Min());
             Assert.Equal(int.MinValue, one.Concat(Enumerable.Repeat(int.MinValue, 1)).Min());
             Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Min());
-        }
-
-        [Fact]
-        public void MaxInt32()
-        {
-            var ten = Enumerable.Range(1, 10).ToArray();
-            var minusTen = new[] { -100, -15, -50, -10 };
-            var thousand = new[] { -16, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat(42, 1).Max());
-            Assert.Equal(10, ten.Max());
-            Assert.Equal(-10, minusTen.Max());
-            Assert.Equal(1000, thousand.Max());
-            Assert.Equal(int.MaxValue, thousand.Concat(Enumerable.Repeat(int.MaxValue, 1)).Max());
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Max());
         }
 
         [Fact]
@@ -51,20 +37,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxInt64()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (long)i).ToArray();
-            var minusTen = new[] { -100L, -15, -50, -10 };
-            var thousand = new[] { -16L, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat(42L, 1).Max());
-            Assert.Equal(10, ten.Max());
-            Assert.Equal(-10, minusTen.Max());
-            Assert.Equal(1000, thousand.Max());
-            Assert.Equal(long.MaxValue, thousand.Concat(Enumerable.Repeat(long.MaxValue, 1)).Max());
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<long>().Max());
-        }
-
-        [Fact]
         public void MinSingle()
         {
             var one = Enumerable.Range(1, 10).Select(i => (float)i).ToArray();
@@ -76,20 +48,6 @@ namespace System.Linq.Tests
             Assert.Equal(100L, hundred.Min());
             Assert.Equal(float.MinValue, one.Concat(Enumerable.Repeat(float.MinValue, 1)).Min());
             Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<float>().Min());
-        }
-
-        [Fact]
-        public void MaxSingle()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (float)i).ToArray();
-            var minusTen = new[] { -100F, -15, -50, -10 };
-            var thousand = new[] { -16F, 0, 50, 100, 1000 };
-            Assert.Equal(42F, Enumerable.Repeat(42F, 1).Max());
-            Assert.Equal(10F, ten.Max());
-            Assert.Equal(-10F, minusTen.Max());
-            Assert.Equal(1000F, thousand.Max());
-            Assert.Equal(float.MaxValue, thousand.Concat(Enumerable.Repeat(float.MaxValue, 1)).Max());
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<float>().Max());
         }
 
         [Fact]
@@ -107,45 +65,17 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxDouble()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (double)i).ToArray();
-            var minusTen = new[] { -100D, -15, -50, -10 };
-            var thousand = new[] { -16D, 0, 50, 100, 1000 };
-            Assert.Equal(42D, Enumerable.Repeat(42D, 1).Max());
-            Assert.Equal(10D, ten.Max());
-            Assert.Equal(-10D, minusTen.Max());
-            Assert.Equal(1000D, thousand.Max());
-            Assert.Equal(double.MaxValue, thousand.Concat(Enumerable.Repeat(double.MaxValue, 1)).Max());
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<double>().Max());
-        }
-
-        [Fact]
         public void MinDecimal()
         {
             var one = Enumerable.Range(1, 10).Select(i => (decimal)i).ToArray();
             var minusTen = new[] { -1M, -10, 10, 200, 1000 };
             var hundred = new[] { 3000M, 100, 200, 1000 };
-            Assert.Equal(42M, Enumerable.Repeat(42M, 1).Max());
+            Assert.Equal(42M, Enumerable.Repeat(42M, 1).Min());
             Assert.Equal(1M, one.Min());
             Assert.Equal(-10M, minusTen.Min());
             Assert.Equal(100M, hundred.Min());
             Assert.Equal(decimal.MinValue, one.Concat(Enumerable.Repeat(decimal.MinValue, 1)).Min());
             Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<decimal>().Min());
-        }
-
-        [Fact]
-        public void MaxDecimal()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (decimal)i).ToArray();
-            var minusTen = new[] { -100M, -15, -50, -10 };
-            var thousand = new[] { -16M, 0, 50, 100, 1000 };
-            Assert.Equal(42M, Enumerable.Repeat(42M, 1).Max());
-            Assert.Equal(10M, ten.Max());
-            Assert.Equal(-10M, minusTen.Max());
-            Assert.Equal(1000M, thousand.Max());
-            Assert.Equal(decimal.MaxValue, thousand.Concat(Enumerable.Repeat(decimal.MaxValue, 1)).Max());
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<decimal>().Max());
         }
 
         [Fact]
@@ -164,21 +94,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxNullableInt32()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (int?)i).ToArray();
-            var minusTen = new[] { default(int?), -100, -15, -50, -10 };
-            var thousand = new[] { default(int?), -16, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat((int?)42, 1).Max());
-            Assert.Equal(10, ten.Max());
-            Assert.Equal(-10, minusTen.Max());
-            Assert.Equal(1000, thousand.Max());
-            Assert.Equal(int.MaxValue, thousand.Concat(Enumerable.Repeat((int?)int.MaxValue, 1)).Max());
-            Assert.Equal(null, Enumerable.Empty<int?>().Max());
-            Assert.Equal(null, Enumerable.Repeat(default(int?), 100).Max());
-        }
-
-        [Fact]
         public void MinNullableInt64()
         {
             var one = Enumerable.Range(1, 10).Select(i => (long?)i).ToArray();
@@ -191,21 +106,6 @@ namespace System.Linq.Tests
             Assert.Equal(null, Enumerable.Empty<long?>().Min());
             Assert.Equal(long.MinValue, one.Concat(Enumerable.Repeat((long?)long.MinValue, 1)).Min());
             Assert.Equal(null, Enumerable.Repeat(default(long?), 100).Min());
-        }
-
-        [Fact]
-        public void MaxNullableInt64()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (long?)i).ToArray();
-            var minusTen = new[] { default(long?), -100L, -15, -50, -10 };
-            var thousand = new[] { default(long?), -16L, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat((long?)42, 1).Max());
-            Assert.Equal(10, ten.Max());
-            Assert.Equal(-10, minusTen.Max());
-            Assert.Equal(1000, thousand.Max());
-            Assert.Equal(long.MaxValue, thousand.Concat(Enumerable.Repeat((long?)long.MaxValue, 1)).Max());
-            Assert.Equal(null, Enumerable.Empty<long?>().Max());
-            Assert.Equal(null, Enumerable.Repeat(default(long?), 100).Max());
         }
 
         [Fact]
@@ -224,21 +124,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxNullableSingle()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (float?)i).ToArray();
-            var minusTen = new[] { default(float?), -100F, -15, -50, -10 };
-            var thousand = new[] { default(float?), -16F, 0, 50, 100, 1000 };
-            Assert.Equal(42F, Enumerable.Repeat((float?)42, 1).Max());
-            Assert.Equal(10F, ten.Max());
-            Assert.Equal(-10F, minusTen.Max());
-            Assert.Equal(1000F, thousand.Max());
-            Assert.Equal(float.MaxValue, thousand.Concat(Enumerable.Repeat((float?)float.MaxValue, 1)).Max());
-            Assert.Equal(null, Enumerable.Empty<float?>().Max());
-            Assert.Equal(null, Enumerable.Repeat(default(float?), 100).Max());
-        }
-
-        [Fact]
         public void MinNullableDouble()
         {
             var one = Enumerable.Range(1, 10).Select(i => (double?)i).ToArray();
@@ -254,21 +139,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxNullableDouble()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (double?)i).ToArray();
-            var minusTen = new[] { default(double?), -100D, -15, -50, -10 };
-            var thousand = new[] { default(double?), -16D, 0, 50, 100, 1000 };
-            Assert.Equal(42D, Enumerable.Repeat((double?)42, 1).Max());
-            Assert.Equal(10D, ten.Max());
-            Assert.Equal(-10D, minusTen.Max());
-            Assert.Equal(1000D, thousand.Max());
-            Assert.Equal(double.MaxValue, thousand.Concat(Enumerable.Repeat((double?)double.MaxValue, 1)).Max());
-            Assert.Equal(null, Enumerable.Empty<double?>().Max());
-            Assert.Equal(null, Enumerable.Repeat(default(double?), 100).Max());
-        }
-
-        [Fact]
         public void MinNullableDecimal()
         {
             var one = Enumerable.Range(1, 10).Select(i => (decimal?)i).ToArray();
@@ -281,21 +151,6 @@ namespace System.Linq.Tests
             Assert.Equal(decimal.MinValue, one.Concat(Enumerable.Repeat((decimal?)decimal.MinValue, 1)).Min());
             Assert.Equal(null, Enumerable.Empty<decimal?>().Min());
             Assert.Equal(null, Enumerable.Repeat(default(decimal?), 100).Min());
-        }
-
-        [Fact]
-        public void MaxNullableDecimal()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (decimal?)i).ToArray();
-            var minusTen = new[] { default(decimal?), -100M, -15, -50, -10 };
-            var thousand = new[] { default(decimal?), -16M, 0, 50, 100, 1000 };
-            Assert.Equal(42M, Enumerable.Repeat((decimal?)42, 1).Max());
-            Assert.Equal(10M, ten.Max());
-            Assert.Equal(-10M, minusTen.Max());
-            Assert.Equal(1000M, thousand.Max());
-            Assert.Equal(decimal.MaxValue, thousand.Concat(Enumerable.Repeat((decimal?)decimal.MaxValue, 1)).Max());
-            Assert.Equal(null, Enumerable.Empty<decimal?>().Max());
-            Assert.Equal(null, Enumerable.Repeat(default(decimal?), 100).Max());
         }
 
         // Normally NaN < anything is false, as is anything < NaN
@@ -318,12 +173,8 @@ namespace System.Linq.Tests
             Assert.True(float.IsNaN(nanThenOne.Min()));
             Assert.True(float.IsNaN(nanThenMinusTen.Min()));
             Assert.True(float.IsNaN(nanThenMinValue.Min()));
-            Assert.False(float.IsNaN(nanThenOne.Max()));
-            Assert.False(float.IsNaN(nanThenMinusTen.Max()));
-            Assert.False(float.IsNaN(nanThenMinValue.Max()));
             var nanWithNull = new[] { default(float?), float.NaN, default(float?) };
             Assert.True(float.IsNaN(nanWithNull.Min().Value));
-            Assert.True(float.IsNaN(nanWithNull.Max().Value));
         }
 
         [Fact]
@@ -335,12 +186,8 @@ namespace System.Linq.Tests
             Assert.True(double.IsNaN(nanThenOne.Min()));
             Assert.True(double.IsNaN(nanThenMinusTen.Min()));
             Assert.True(double.IsNaN(nanThenMinValue.Min()));
-            Assert.False(double.IsNaN(nanThenOne.Max()));
-            Assert.False(double.IsNaN(nanThenMinusTen.Max()));
-            Assert.False(double.IsNaN(nanThenMinValue.Max()));
             var nanWithNull = new[] { default(double?), double.NaN, default(double?) };
             Assert.True(double.IsNaN(nanWithNull.Min().Value));
-            Assert.True(double.IsNaN(nanWithNull.Max().Value));
         }
 
         [Fact]
@@ -368,30 +215,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxDateTime()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => new DateTime(2000, 1, i)).ToArray();
-            var newYearsEve = new[]
-            {
-                new DateTime(2000, 12, 1),
-                new DateTime(2000, 12, 31),
-                new DateTime(2000, 1, 12)
-            };
-            var threeThousand = new[]
-            {
-                new DateTime(3000, 1, 1),
-                new DateTime(100, 1, 1),
-                new DateTime(200, 1, 1),
-                new DateTime(1000, 1, 1)
-            };
-            Assert.Equal(new DateTime(2000, 1, 10), ten.Max());
-            Assert.Equal(new DateTime(2000, 12, 31), newYearsEve.Max());
-            Assert.Equal(new DateTime(3000, 1, 1), threeThousand.Max());
-            Assert.Equal(DateTime.MaxValue, threeThousand.Concat(Enumerable.Repeat(DateTime.MaxValue, 1)).Max());
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<DateTime>().Max());
-        }
-
-        [Fact]
         public void MinString()
         {
             var one = Enumerable.Range(1, 10).Select(i => i.ToString()).ToArray();
@@ -404,18 +227,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxString()
-        {
-            var nine = Enumerable.Range(1, 10).Select(i => i.ToString()).ToArray();
-            var agents = new[] { "Alice", "Bob", "Charlie", "Eve", "Mallory", "Victor", "Trent" };
-            var confusedAgents = new[] { null, "Charlie", null, "Victor", "Trent", null, "Eve", "Alice", "Mallory", "Bob" };
-            Assert.Equal("9", nine.Max());
-            Assert.Equal("Victor", agents.Max());
-            Assert.Equal("Victor", confusedAgents.Max());
-            Assert.Null(Enumerable.Empty<string>().Max());
-        }
-        [Fact]
-
         public void MinInt32WithSelector()
         {
             var one = Enumerable.Range(1, 10).ToArray();
@@ -427,20 +238,6 @@ namespace System.Linq.Tests
             Assert.Equal(100, hundred.Min(x => x));
             Assert.Equal(int.MinValue, one.Concat(Enumerable.Repeat(int.MinValue, 1)).Min(x => x));
             Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Min(x => x));
-        }
-
-        [Fact]
-        public void MaxInt32WithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).ToArray();
-            var minusTen = new[] { -100, -15, -50, -10 };
-            var thousand = new[] { -16, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat(42, 1).Max(x => x));
-            Assert.Equal(10, ten.Max(x => x));
-            Assert.Equal(-10, minusTen.Max(x => x));
-            Assert.Equal(1000, thousand.Max(x => x));
-            Assert.Equal(int.MaxValue, thousand.Concat(Enumerable.Repeat(int.MaxValue, 1)).Max(x => x));
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<int>().Max(x => x));
         }
 
         [Fact]
@@ -458,20 +255,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxInt64WithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (long)i).ToArray();
-            var minusTen = new[] { -100L, -15, -50, -10 };
-            var thousand = new[] { -16L, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat(42L, 1).Max(x => x));
-            Assert.Equal(10, ten.Max(x => x));
-            Assert.Equal(-10, minusTen.Max(x => x));
-            Assert.Equal(1000, thousand.Max(x => x));
-            Assert.Equal(long.MaxValue, thousand.Concat(Enumerable.Repeat(long.MaxValue, 1)).Max(x => x));
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<long>().Max(x => x));
-        }
-
-        [Fact]
         public void MinSingleWithSelector()
         {
             var one = Enumerable.Range(1, 10).Select(i => (float)i).ToArray();
@@ -483,20 +266,6 @@ namespace System.Linq.Tests
             Assert.Equal(100L, hundred.Min(x => x));
             Assert.Equal(float.MinValue, one.Concat(Enumerable.Repeat(float.MinValue, 1)).Min(x => x));
             Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<float>().Min(x => x));
-        }
-
-        [Fact]
-        public void MaxSingleWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (float)i).ToArray();
-            var minusTen = new[] { -100F, -15, -50, -10 };
-            var thousand = new[] { -16F, 0, 50, 100, 1000 };
-            Assert.Equal(42F, Enumerable.Repeat(42F, 1).Max(x => x));
-            Assert.Equal(10F, ten.Max(x => x));
-            Assert.Equal(-10F, minusTen.Max(x => x));
-            Assert.Equal(1000F, thousand.Max(x => x));
-            Assert.Equal(float.MaxValue, thousand.Concat(Enumerable.Repeat(float.MaxValue, 1)).Max(x => x));
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<float>().Max(x => x));
         }
 
         [Fact]
@@ -514,45 +283,17 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxDoubleWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (double)i).ToArray();
-            var minusTen = new[] { -100D, -15, -50, -10 };
-            var thousand = new[] { -16D, 0, 50, 100, 1000 };
-            Assert.Equal(42D, Enumerable.Repeat(42D, 1).Max(x => x));
-            Assert.Equal(10D, ten.Max(x => x));
-            Assert.Equal(-10D, minusTen.Max(x => x));
-            Assert.Equal(1000D, thousand.Max(x => x));
-            Assert.Equal(double.MaxValue, thousand.Concat(Enumerable.Repeat(double.MaxValue, 1)).Max(x => x));
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<double>().Max(x => x));
-        }
-
-        [Fact]
         public void MinDecimalWithSelector()
         {
             var one = Enumerable.Range(1, 10).Select(i => (decimal)i).ToArray();
             var minusTen = new[] { -1M, -10, 10, 200, 1000 };
             var hundred = new[] { 3000M, 100, 200, 1000 };
-            Assert.Equal(42M, Enumerable.Repeat(42M, 1).Max(x => x));
+            Assert.Equal(42M, Enumerable.Repeat(42M, 1).Min(x => x));
             Assert.Equal(1M, one.Min(x => x));
             Assert.Equal(-10M, minusTen.Min(x => x));
             Assert.Equal(100M, hundred.Min(x => x));
             Assert.Equal(decimal.MinValue, one.Concat(Enumerable.Repeat(decimal.MinValue, 1)).Min(x => x));
             Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<decimal>().Min(x => x));
-        }
-
-        [Fact]
-        public void MaxDecimalWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (decimal)i).ToArray();
-            var minusTen = new[] { -100M, -15, -50, -10 };
-            var thousand = new[] { -16M, 0, 50, 100, 1000 };
-            Assert.Equal(42M, Enumerable.Repeat(42M, 1).Max(x => x));
-            Assert.Equal(10M, ten.Max(x => x));
-            Assert.Equal(-10M, minusTen.Max(x => x));
-            Assert.Equal(1000M, thousand.Max(x => x));
-            Assert.Equal(decimal.MaxValue, thousand.Concat(Enumerable.Repeat(decimal.MaxValue, 1)).Max(x => x));
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<decimal>().Max(x => x));
         }
 
         [Fact]
@@ -571,21 +312,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxNullableInt32WithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (int?)i).ToArray();
-            var minusTen = new[] { default(int?), -100, -15, -50, -10 };
-            var thousand = new[] { default(int?), -16, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat((int?)42, 1).Max(x => x));
-            Assert.Equal(10, ten.Max(x => x));
-            Assert.Equal(-10, minusTen.Max(x => x));
-            Assert.Equal(1000, thousand.Max(x => x));
-            Assert.Equal(int.MaxValue, thousand.Concat(Enumerable.Repeat((int?)int.MaxValue, 1)).Max(x => x));
-            Assert.Equal(null, Enumerable.Empty<int?>().Max(x => x));
-            Assert.Equal(null, Enumerable.Repeat(default(int?), 100).Max(x => x));
-        }
-
-        [Fact]
         public void MinNullableInt64WithSelector()
         {
             var one = Enumerable.Range(1, 10).Select(i => (long?)i).ToArray();
@@ -598,21 +324,6 @@ namespace System.Linq.Tests
             Assert.Equal(null, Enumerable.Empty<long?>().Min(x => x));
             Assert.Equal(long.MinValue, one.Concat(Enumerable.Repeat((long?)long.MinValue, 1)).Min(x => x));
             Assert.Equal(null, Enumerable.Repeat(default(long?), 100).Min(x => x));
-        }
-
-        [Fact]
-        public void MaxNullableInt64WithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (long?)i).ToArray();
-            var minusTen = new[] { default(long?), -100L, -15, -50, -10 };
-            var thousand = new[] { default(long?), -16L, 0, 50, 100, 1000 };
-            Assert.Equal(42, Enumerable.Repeat((long?)42, 1).Max(x => x));
-            Assert.Equal(10, ten.Max(x => x));
-            Assert.Equal(-10, minusTen.Max(x => x));
-            Assert.Equal(1000, thousand.Max(x => x));
-            Assert.Equal(long.MaxValue, thousand.Concat(Enumerable.Repeat((long?)long.MaxValue, 1)).Max(x => x));
-            Assert.Equal(null, Enumerable.Empty<long?>().Max(x => x));
-            Assert.Equal(null, Enumerable.Repeat(default(long?), 100).Max(x => x));
         }
 
         [Fact]
@@ -631,21 +342,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxNullableSingleWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (float?)i).ToArray();
-            var minusTen = new[] { default(float?), -100F, -15, -50, -10 };
-            var thousand = new[] { default(float?), -16F, 0, 50, 100, 1000 };
-            Assert.Equal(42F, Enumerable.Repeat((float?)42, 1).Max(x => x));
-            Assert.Equal(10F, ten.Max(x => x));
-            Assert.Equal(-10F, minusTen.Max(x => x));
-            Assert.Equal(1000F, thousand.Max(x => x));
-            Assert.Equal(float.MaxValue, thousand.Concat(Enumerable.Repeat((float?)float.MaxValue, 1)).Max(x => x));
-            Assert.Equal(null, Enumerable.Empty<float?>().Max(x => x));
-            Assert.Equal(null, Enumerable.Repeat(default(float?), 100).Max(x => x));
-        }
-
-        [Fact]
         public void MinNullableDoubleWithSelector()
         {
             var one = Enumerable.Range(1, 10).Select(i => (double?)i).ToArray();
@@ -661,21 +357,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxNullableDoubleWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (double?)i).ToArray();
-            var minusTen = new[] { default(double?), -100D, -15, -50, -10 };
-            var thousand = new[] { default(double?), -16D, 0, 50, 100, 1000 };
-            Assert.Equal(42D, Enumerable.Repeat((double?)42, 1).Max(x => x));
-            Assert.Equal(10D, ten.Max(x => x));
-            Assert.Equal(-10D, minusTen.Max(x => x));
-            Assert.Equal(1000D, thousand.Max(x => x));
-            Assert.Equal(double.MaxValue, thousand.Concat(Enumerable.Repeat((double?)double.MaxValue, 1)).Max(x => x));
-            Assert.Equal(null, Enumerable.Empty<double?>().Max(x => x));
-            Assert.Equal(null, Enumerable.Repeat(default(double?), 100).Max(x => x));
-        }
-
-        [Fact]
         public void MinNullableDecimalWithSelector()
         {
             var one = Enumerable.Range(1, 10).Select(i => (decimal?)i).ToArray();
@@ -688,21 +369,6 @@ namespace System.Linq.Tests
             Assert.Equal(decimal.MinValue, one.Concat(Enumerable.Repeat((decimal?)decimal.MinValue, 1)).Min(x => x));
             Assert.Equal(null, Enumerable.Empty<decimal?>().Min(x => x));
             Assert.Equal(null, Enumerable.Repeat(default(decimal?), 100).Min(x => x));
-        }
-
-        [Fact]
-        public void MaxNullableDecimalWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => (decimal?)i).ToArray();
-            var minusTen = new[] { default(decimal?), -100M, -15, -50, -10 };
-            var thousand = new[] { default(decimal?), -16M, 0, 50, 100, 1000 };
-            Assert.Equal(42M, Enumerable.Repeat((decimal?)42, 1).Max(x => x));
-            Assert.Equal(10M, ten.Max(x => x));
-            Assert.Equal(-10M, minusTen.Max(x => x));
-            Assert.Equal(1000M, thousand.Max(x => x));
-            Assert.Equal(decimal.MaxValue, thousand.Concat(Enumerable.Repeat((decimal?)decimal.MaxValue, 1)).Max(x => x));
-            Assert.Equal(null, Enumerable.Empty<decimal?>().Max(x => x));
-            Assert.Equal(null, Enumerable.Repeat(default(decimal?), 100).Max(x => x));
         }
 
         // Normally NaN < anything is false, as is anything < NaN
@@ -725,12 +391,8 @@ namespace System.Linq.Tests
             Assert.True(float.IsNaN(nanThenOne.Min(x => x)));
             Assert.True(float.IsNaN(nanThenMinusTen.Min(x => x)));
             Assert.True(float.IsNaN(nanThenMinValue.Min(x => x)));
-            Assert.False(float.IsNaN(nanThenOne.Max(x => x)));
-            Assert.False(float.IsNaN(nanThenMinusTen.Max(x => x)));
-            Assert.False(float.IsNaN(nanThenMinValue.Max(x => x)));
             var nanWithNull = new[] { default(float?), float.NaN, default(float?) };
             Assert.True(float.IsNaN(nanWithNull.Min(x => x).Value));
-            Assert.True(float.IsNaN(nanWithNull.Max(x => x).Value));
         }
 
         [Fact]
@@ -742,12 +404,8 @@ namespace System.Linq.Tests
             Assert.True(double.IsNaN(nanThenOne.Min(x => x)));
             Assert.True(double.IsNaN(nanThenMinusTen.Min(x => x)));
             Assert.True(double.IsNaN(nanThenMinValue.Min(x => x)));
-            Assert.False(double.IsNaN(nanThenOne.Max(x => x)));
-            Assert.False(double.IsNaN(nanThenMinusTen.Max(x => x)));
-            Assert.False(double.IsNaN(nanThenMinValue.Max(x => x)));
             var nanWithNull = new[] { default(double?), double.NaN, default(double?) };
             Assert.True(double.IsNaN(nanWithNull.Min(x => x).Value));
-            Assert.True(double.IsNaN(nanWithNull.Max(x => x).Value));
         }
 
         [Fact]
@@ -775,30 +433,6 @@ namespace System.Linq.Tests
         }
 
         [Fact]
-        public void MaxDateTimeWithSelector()
-        {
-            var ten = Enumerable.Range(1, 10).Select(i => new DateTime(2000, 1, i)).ToArray();
-            var newYearsEve = new[]
-            {
-                new DateTime(2000, 12, 1),
-                new DateTime(2000, 12, 31),
-                new DateTime(2000, 1, 12)
-            };
-            var threeThousand = new[]
-            {
-                new DateTime(3000, 1, 1),
-                new DateTime(100, 1, 1),
-                new DateTime(200, 1, 1),
-                new DateTime(1000, 1, 1)
-            };
-            Assert.Equal(new DateTime(2000, 1, 10), ten.Max(x => x));
-            Assert.Equal(new DateTime(2000, 12, 31), newYearsEve.Max(x => x));
-            Assert.Equal(new DateTime(3000, 1, 1), threeThousand.Max(x => x));
-            Assert.Equal(DateTime.MaxValue, threeThousand.Concat(Enumerable.Repeat(DateTime.MaxValue, 1)).Max(x => x));
-            Assert.Throws<InvalidOperationException>(() => Enumerable.Empty<DateTime>().Max(x => x));
-        }
-
-        [Fact]
         public void MinStringWithSelector()
         {
             var one = Enumerable.Range(1, 10).Select(i => i.ToString()).ToArray();
@@ -808,18 +442,6 @@ namespace System.Linq.Tests
             Assert.Equal("Alice", agents.Min(x => x));
             Assert.Equal("Alice", confusedAgents.Min(x => x));
             Assert.Null(Enumerable.Empty<string>().Min(x => x));
-        }
-
-        [Fact]
-        public void MaxStringWithSelector()
-        {
-            var nine = Enumerable.Range(1, 10).Select(i => i.ToString()).ToArray();
-            var agents = new[] { "Alice", "Bob", "Charlie", "Eve", "Mallory", "Victor", "Trent" };
-            var confusedAgents = new[] { null, "Charlie", null, "Victor", "Trent", null, "Eve", "Alice", "Mallory", "Bob" };
-            Assert.Equal("9", nine.Max(x => x));
-            Assert.Equal("Victor", agents.Max(x => x));
-            Assert.Equal("Victor", confusedAgents.Max(x => x));
-            Assert.Null(Enumerable.Empty<string>().Max(x => x));
         }
     }
 }
