+++ src/System.Linq/tests/System.Linq.Tests.csproj
@@ -35,7 +35,8 @@
     <Compile Include="IntersectTests.cs" />
     <Compile Include="LastOrDefaultTests.cs" />
     <Compile Include="LastTests.cs" />
-    <Compile Include="MinMaxTests.cs" />
+    <Compile Include="MaxTests.cs" />
+    <Compile Include="MinTests.cs" />
     <Compile Include="Performance\CastPerf.cs" />
     <Compile Include="Performance\LinqPerformanceCore.cs" />
     <Compile Include="Performance\OrderByThenByPerf.cs" />
